# api/openapi.yaml
openapi: 3.1.0
info:
  title: FlowPilot API
  version: 0.1.0
  description: >
    Contract-first REST API for FlowPilot. Paths and schemas are added slice
    by slice. PR1 adds Auth, PR2 adds the read-only user directory, PR3 adds
    Projects + BoardColumns, PR4 adds project Members, PR5 adds WorkItems.
    Hand-authored here is the source of truth; springdoc-openapi runs in CI
    only as a drift check against the running application (see
    OpenApiSpecExportTest + the oasdiff step in ci.yml). Paths below are
    absolute (include the /api prefix baked into every @RequestMapping) —
    there is no servers/context-path indirection, since the live
    springdoc-generated document has none either.
paths:
...
  /api/projects/{projectId}/work-items:
    get:
      summary: List a project's work items
      operationId: listWorkItems
      description: >-
        Requires the same admin/owner/member view gate as GET /projects/{id}.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NestedProjectId'
      responses:
        '200':
          description: Work items, ordered by column then within-column position
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkItemResponse'
        '403':
          description: Caller is not the owner, an administrator, or a member
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    post:
      summary: Create a work item
      operationId: createWorkItem
      description: >-
        Any project member (owner, a global administrator, or a live
        ProjectMember of the project) may create a work item — task
        management must stay usable by the whole team, not just the project
        owner. The item is always created in the project's first
        (lowest-position) BoardColumn; moving between columns is a separate
        slice-6 endpoint.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NestedProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkItemCreateRequest'
      responses:
        '201':
          description: Work item created in the project's first column
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemResponse'
        '403':
          description: Caller is not the owner, an administrator, or a member
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: No project with that id
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/work-items/{id}:
    get:
      summary: Get a work item by id
      operationId: getWorkItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkItemId'
      responses:
        '200':
          description: Work item found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemResponse'
        '403':
          description: Caller is not the owner, an administrator, or a member of the parent project
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: No work item with that id
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    put:
      summary: Update a work item's title/description/assignee
      operationId: updateWorkItem
      description: >-
        Any project member (owner, a global administrator, or a live
        ProjectMember of the project) may edit. Does NOT accept
        columnId/position — moving is the separate slice-6
        PUT /work-items/{id}/move endpoint.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkItemUpdateRequest'
      responses:
        '200':
          description: Work item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemResponse'
        '403':
          description: Caller is not the owner, an administrator, or a member
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: No work item with that id
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    delete:
      summary: Delete a work item
      operationId: deleteWorkItem
      description: >-
        Any project member (owner, a global administrator, or a live
        ProjectMember of the project) may delete.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkItemId'
      responses:
        '204':
          description: Work item deleted
        '403':
          description: Caller is not the owner, an administrator, or a member
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: No work item with that id
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/work-items/{id}/move:
    put:
      summary: Move a work item to a column and/or position
      operationId: moveWorkItem
      description: >-
        Any project member (owner, a global administrator, or a live
        ProjectMember of the project) may move. Validates that columnId
        belongs to the SAME project as the work item; a cross-project column
        is rejected with 400 and no change. position is the desired
        zero-based insertion index within the target column's existing
        items (the moved item itself excluded), translated server-side into
        a gap-based position (1024 step; the target column is fully
        re-sequenced only when the computed gap would be less than 2).
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkItemMoveRequest'
      responses:
        '200':
          description: >-
            Work item moved; columnId and position persisted. affectedItems
            is populated only when the target column's siblings were
            re-sequenced.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemResponse'
        '400':
          description: columnId belongs to a different project
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Caller is not the owner, an administrator, or a member
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: No work item or no target column with that id
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
components:
  parameters:
...
    NestedProjectId:
      name: projectId
      in: path
      required: true
      description: >-
        Same conceptual project id as ProjectId, but named projectId here
        because that's the literal @RequestMapping path-variable name on
        ProjectMemberController/WorkItemController's nested-under-project
        routes (unlike ProjectController's own routes, which use {id}).
      schema:
        type: integer
        format: int64
...
    WorkItemId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
    WorkItemPriority:
      type: string
      enum: [LOW, MEDIUM, HIGH, URGENT]
...
    WorkItemCreateRequest:
      type: object
      description: >-
        No columnId — the item always lands in the project's first seeded
        BoardColumn (spec: work-items).
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        assignedUserId:
          type: integer
          format: int64
        sprintId:
          type: [integer, 'null']
          format: int64
        priority:
          $ref: '#/components/schemas/WorkItemPriority'
    WorkItemUpdateRequest:
      type: object
      description: >-
        Excludes columnId/position — those are mutated only by the slice-6
        move endpoint.
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        assignedUserId:
          type: integer
          format: int64
        sprintId:
          type: [integer, 'null']
          format: int64
        priority:
          $ref: '#/components/schemas/WorkItemPriority'
    WorkItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        projectId:
          type: integer
          format: int64
        columnId:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        assignedUserId:
          type: integer
          format: int64
        assignedUserName:
          type: string
          nullable: true
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        sprintId:
          type: integer
          format: int64
          nullable: true
        priority:
          $ref: '#/components/schemas/WorkItemPriority'
        affectedItems:
          type: array
          nullable: true
          description: >-
            Populated only by PUT /api/work-items/{id}/move when the target
            column's siblings were re-sequenced (design D10); null/absent
            otherwise, including for every non-move response.
          items:
            $ref: '#/components/schemas/WorkItemResponse'
    WorkItemMoveRequest:
      type: object
      description: >-
        position is the desired zero-based insertion index within the
        target column's existing items (the moved item itself excluded),
        not a raw gap-based position value.
      required: [columnId, position]
      properties:
        columnId:
          type: integer
          format: int64
        position:
          type: integer
          format: int32
          minimum: 0
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT